requirements.txt
# Core dependencies
wheel
PyQt5>=5.15.0
numpy>=1.19.0
matplotlib>=3.3.0

# Machine learning
torch>=1.9.0
torchvision>=0.10.0

# Chess-specific libraries
python-chess>=1.0.0

# Utilities
tqdm>=4.50.0
pandas>=1.1.0

# For Windows systems - GMPY2 wheels
https://download.lfd.uci.edu/pythonlibs/z2tqcw5k/gmpy2-2.0.8-cp38-cp38-win_amd64.whl ; platform_system=="Windows" and python_version=="3.8"
https://download.lfd.uci.edu/pythonlibs/z2tqcw5k/gmpy2-2.0.8-cp39-cp39-win_amd64.whl ; platform_system=="Windows" and python_version=="3.9"
https://download.lfd.uci.edu/pythonlibs/z2tqcw5k/gmpy2-2.0.8-cp310-cp310-win_amd64.whl ; platform_system=="Windows" and python_version=="3.10"
gmpy2 ; platform_system=="Linux" or platform_system=="Darwin"